+++ include/Vector2D.hpp
@@ -2,10 +2,10 @@
 
 class Vector2D {
 public:
-    float x, y;
+    double x, y;
 
     // Constructor
-    Vector2D(float x = 0.0f, float y = 0.0f);
+    Vector2D(double x = 0.0f, double y = 0.0f);
 
     // Operator Overloads
     Vector2D operator+(const Vector2D& rhs) const;
